Move resume page types to views directory
